Slight modifications to conform to some standards
